+++ src/main/java/com/thoughtclan/bikerentalsystem/repositories/PrivilegesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PrivilegesRepository extends JpaRepository<privileges, Integer> {
+public interface PrivilegesRepository extends JpaRepository<privileges, Long> {
 }
